move ComponentRequestHandler into streamlit.web.server
